Handle change in default type of dates
With the object conversion path refacter, Arrow now returns the same
date type when created from a list and a numpy array. Sadly this is
different to what we use as the reference type in our tests. To fix the
root of the problem, the Pandas tests should be adjusted to pass in the
dtype in their calls to `_from_sequence`.